fix COHD client; add max_results back
